+++ index.js
@@ -22,8 +22,19 @@ function IMF (opts) {
     if (opts.languages || opts.callback) {
         this.loadLocales(opts.languages, function () {
             var locales = Array.from(arguments);
-            if (opts.callback) {
-                opts.callback.apply(opts.callback, [that.getFormatter(opts.namespace), that.getFormatter.bind(that), locales]);
+            function runCallback () {
+                if (opts.callback) {
+                    opts.callback.apply(opts.callback, [that.getFormatter(opts.namespace), that.getFormatter.bind(that), locales]);
+                }
+            }
+            if (opts.fallbackLanguage) {
+                that.loadLocales(opts.fallbackLanguage, function (fallbackLocale) {
+                    that.fallbackLocale = fallbackLocale;
+                    runCallback();
+                }, true);
+            }
+            else {
+                runCallback();
             }
         });
     }
@@ -44,13 +55,23 @@ IMF.prototype.getFormatter = function (ns, sep) {
     sep = sep === undefined ? this.defaultSeparator : sep;
     ns = Array.isArray(ns) ? ns.join(sep) : ns;
 
-    return function (key, values, formats) {
+    return function (key, values, formats, fallback) {
         var message;
-        that.locales.some(function (locale) {
+        if (key && typeof key === 'object') {
+            values = key.values;
+            formats = key.formats;
+            fallback = key.fallback;
+            key = key.key;
+        }
+        function findMessage (locale) {
             message = locale[(ns ? ns + sep : '') + key] || messageForNSParts(locale, ns, sep, key);
             return message;
-        });
+        }
+        that.locales.some(findMessage);
         if (!message) {
+            if (fallback) {
+                return fallback({message: that.fallbackLocale && findMessage(that.fallbackLocale), langs: that.langs, namespace: ns, separator: sep, key: key, values: values, formats: formats});
+            }
             throw "Message not found for locales " + that.langs + " with key " + key + ", namespace " + ns + ", and namespace separator " + sep;
         }
         if (!values && !formats) {
@@ -61,16 +82,19 @@ IMF.prototype.getFormatter = function (ns, sep) {
     };
 };
 
-IMF.prototype.loadLocales = function (langs, cb) {
+IMF.prototype.loadLocales = function (langs, cb, avoidSettingLocales) {
     var that = this;
     langs = langs || navigator.language || 'en-US';
     this.langs = Array.isArray(langs) ? langs : [langs];
     getJSON(
         this.langs.map(this.localeFileResolver, this),
         function () {
-            that.locales = Array.from(arguments);
+            var locales = Array.from(arguments);
+            if (!avoidSettingLocales) {
+                that.locales = locales;
+            }
             if (cb) {
-                cb.apply(that, that.locales);
+                cb.apply(that, locales);
             }
         }
     );